+++ funsor/__init__.py
@@ -3,9 +3,9 @@ from __future__ import absolute_import, division, print_function
 from funsor.domains import Domain, bint, find_domain, reals
 from funsor.interpreter import reinterpret
 from funsor.terms import Funsor, Number, Variable, of_shape, to_funsor
-from funsor.torch import Function, Tensor, arange, einsum, function
+from funsor.torch import Function, Tensor, arange, torch_einsum, function
 
-from . import distributions, domains, gaussian, handlers, interpreter, minipyro, ops, terms, torch
+from . import distributions, domains, einsum, gaussian, handlers, interpreter, minipyro, ops, terms, torch
 
 __all__ = [
     'Domain',
@@ -33,4 +33,5 @@ __all__ = [
     'terms',
     'to_funsor',
     'torch',
+    'torch_einsum',
 ]
