+++ funsor/testing.py
@@ -9,10 +9,9 @@ import pytest
 import torch
 from six.moves import reduce
 
-import funsor.ops as ops
 from funsor.domains import Domain, bint
 from funsor.gaussian import Gaussian
-from funsor.terms import Binary, Funsor
+from funsor.terms import Funsor
 from funsor.torch import Tensor
 
 
@@ -140,27 +139,3 @@ def random_gaussian(inputs):
     prec_sqrt = torch.randn(batch_shape + event_shape + event_shape)
     precision = torch.matmul(prec_sqrt, prec_sqrt.transpose(-1, -2))
     return Gaussian(log_density, loc, precision, inputs)
-
-
-def naive_einsum(eqn, *terms, **kwargs):
-    backend = kwargs.pop('backend', 'torch')
-    if backend == 'torch':
-        sum_op, prod_op = ops.add, ops.mul
-    elif backend == 'pyro.ops.einsum.torch_log':
-        sum_op, prod_op = ops.logaddexp, ops.add
-    else:
-        raise ValueError("{} backend not implemented".format(backend))
-
-    assert isinstance(eqn, str)
-    assert all(isinstance(term, Funsor) for term in terms)
-    inputs, output = eqn.split('->')
-    assert len(output.split(',')) == 1
-    input_dims = frozenset(d for inp in inputs.split(',') for d in inp)
-    output_dims = frozenset(d for d in output)
-    reduce_dims = tuple(d for d in input_dims - output_dims)
-    prod = terms[0]
-    for term in terms[1:]:
-        prod = Binary(prod_op, prod, term)
-    for reduce_dim in reduce_dims:
-        prod = prod.reduce(sum_op, reduce_dim)
-    return prod
